+++ Tracer/src/shared/hittableList.cxx
@@ -1,6 +1,6 @@
 #include "hittableList.hpp"
 
-bool HittableList::hit(const ray& r, double tMin, double tMax, HitRecord& rec) const {
+bool HittableList::hit(const ray& r, Float tMin, Float tMax, HitRecord& rec) const {
     HitRecord tempRecord;
     bool hitAnything = false;
     auto closestSoFar = tMax;
